setup(AuthContext): Dedicated types folder/file for context
